fix: disable LastUpdateField if automated
