+++ src/lib/components/Form/Field/LastUpdatedField_11.svelte
@@ -6,107 +6,38 @@
   import { invalidateAll } from "$app/navigation";
   import DateInput from "../Inputs/DateInput.svelte";
   import type { MaintenanceFrequency } from "$lib/models/metadata";
+  import { getLastUpdateValue } from "../../../util";
 
   const KEY = 'isoMetadata.modified';
   const LABEL = 'letzte Aktualisierung';
 
-  const getLastUpdateValue = (published: string, maintenanceFrequency: MaintenanceFrequency) => {
-    const publishedDate = new Date(published);
-    const todayDate = new Date();
-    let updateDate = new Date(published);
-
-    function dayInYearIsGreater(dateA: Date, dateB: Date): boolean {
-      const todayYearDay = dateB.getMonth() * 31 + dateB.getDate();
-      const publishedYearDay = dateA.getMonth() * 31 + dateA.getDate();
-      return todayYearDay > publishedYearDay;
-    }
-
-    function getLastUpdateDateMonthly(publishedDate: Date, cycleMonths: number) {
-      const today = new Date();
-      const diffInMonths = (today.getFullYear() - publishedDate.getFullYear()) * 12 + (today.getMonth() - publishedDate.getMonth());
-      const cycles = Math.floor(diffInMonths / cycleMonths);
-
-      const lastUpdate = new Date(publishedDate);
-      lastUpdate.setMonth(lastUpdate.getMonth() + cycles * cycleMonths);
-      return lastUpdate;
-    }
-
-    function getLastUpdateDateDaily(publishedDate: Date, days: number = 14) {
-      const today = new Date();
-      const diffInTime = today.getTime() - publishedDate.getTime();
-      const diffInDays = Math.floor(diffInTime / (1000 * 60 * 60 * 24));
-      const cycles = Math.floor(diffInDays / days);
-
-      const lastUpdate = new Date(publishedDate);
-      lastUpdate.setDate(lastUpdate.getDate() + cycles * days);
-      return lastUpdate;
-    }
-
-    switch (maintenanceFrequency) {
-    case 'continual':
-      updateDate = todayDate;
-      break;
-    case 'daily':
-      updateDate = todayDate;
-      updateDate.setDate(updateDate.getDate() - 1);
-      break;
-    case 'weekly':
-      updateDate = todayDate;
-      if (todayDate.getDay() > publishedDate.getDay()) {
-        updateDate.setDate(todayDate.getDate() - (todayDate.getDay() - publishedDate.getDay()));
-      } else {
-        updateDate.setDate(todayDate.getDate() - (7 - (publishedDate.getDay() - todayDate.getDay())));
-      }
-      break;
-    case 'fortnightly':
-      updateDate = getLastUpdateDateDaily(publishedDate, 14);
-      break;
-    case 'monthly':
-      updateDate = todayDate;
-      if (todayDate.getDate() > publishedDate.getDate()) {
-        updateDate.setDate(publishedDate.getDate());
-      } else {
-        updateDate.setMonth(updateDate.getMonth() - 1);
-        updateDate.setDate(publishedDate.getDate());
-      }
-      break;
-    case 'quarterly':
-      updateDate = getLastUpdateDateMonthly(publishedDate, 3);
-      break;
-    case 'biannually':
-      updateDate = getLastUpdateDateMonthly(publishedDate, 6);
-      break;
-    case 'annually':
-      updateDate = publishedDate;
-      if (dayInYearIsGreater(publishedDate, todayDate)) {
-        updateDate.setFullYear(todayDate.getFullYear());
-      } else {
-        updateDate.setFullYear(todayDate.getFullYear() - 1);
-      }
-      break;
-    case 'asNeeded':
-    case 'irregular':
-    case 'notPlanned':
-    case 'unknown':
-    default:
-      return undefined;
-    }
-    return updateDate;
-  };
-
-  let initialValue = getValue<string>(KEY);
+  const initialValue = getValue<string>(KEY);
   let publishedValue = getValue<string>('isoMetadata.published');
   let maintenanceFrequencyValue = getValue<MaintenanceFrequency>('isoMetadata.maintenanceFrequency');
-
-  if (initialValue) {
-    initialValue = new Date(initialValue).toISOString().split('T')[0];
-  } else if (publishedValue && maintenanceFrequencyValue) {
-    const lastUpdateDate = getLastUpdateValue(publishedValue, maintenanceFrequencyValue);
-    initialValue = lastUpdateDate?.toISOString().split('T')[0] || '';
-  }
-
   let value = $state(initialValue || '');
+
+  // TODO: check why this is not updated in real time
+  $effect(() => {
+    if (initialValue) {
+      value = new Date(initialValue).toISOString().split('T')[0];
+    } else if (publishedValue && maintenanceFrequencyValue) {
+      const lastUpdateDate = getLastUpdateValue(publishedValue, maintenanceFrequencyValue);
+      value = lastUpdateDate?.toISOString().split('T')[0] || '';
+    }
+  });
+
   let showCheckmark = $state(false);
+  let isAutomatedValue = $derived([
+    'continual',
+    'daily',
+    'weekly',
+    'fortnightly',
+    'monthly',
+    'quarterly',
+    'biannually',
+    'annually'
+  ].includes(maintenanceFrequencyValue!));
+  let readOnly = $derived(!!publishedValue && isAutomatedValue);
 
   const onBlur = async () => {
     // TODO check if value has changed
@@ -134,7 +65,7 @@
       key={KEY}
       label={LABEL}
       onblur={onBlur}
-      disabled={!!(publishedValue && maintenanceFrequencyValue)}
+      disabled={readOnly}
     />
   </Paper>
   <FieldTools
